add: OpenApi spec for routing with builtin swagger-ui

+++ solution/src/forms/time/form.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, ToSchema)]
 pub struct TimeForm {
-    pub current_date: Option<i64>,
+    pub current_date: Option<u8>,
 }
