+++ solution/src/models/time/mod.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, ToSchema)]
 pub struct TimeModel {
-    pub current_date: i32,
+    pub current_date: u8,
 }
